requirements.txt
## Minimal core requirements for running the Streamlit dashboard on Streamlit Community Cloud
## Keep heavy ML / AI packages in separate extras files (requirements-ml.txt, requirements-ai.txt)

# Web / dashboard
streamlit==1.19.0

# Data & plotting
numpy>=1.26.3,<3.0
pandas>=2.3.3,<3.1
plotly>=5.15.0
altair==4.2.2
pycountry
python-dotenv>=1.0

# Optional mapping / visualization helpers (kept small)
pydeck>=0.9.1
folium>=0.20.0
branca>=0.8.2
xyzservices

# Pin protobuf to avoid v4 breaking changes
protobuf==3.20.3

# Utilities
tqdm>=4.67.1
validators>=0.35.0

# Note: If you need ML/AI features (model training, RAG), install the extras files:
#  pip install -r requirements-ml.txt
#  pip install -r requirements-ai.txt
